Implemented floating dollar signs for when money is received (along with new C-side class THSpriteRenderList for implementing it). Added left/right/centre alignment option to THFont::drawWrapped, which the fax dialog now uses. Removed iAddedRowDistance argument from THFont::drawWrapped as it duplicated the behaviour of the font's y-separation. Fixed a crash which could occur after a depersistance operation failed (due to leaving a userdata in a partially initialised state). Updated THLinkList and THDrawable structures to use the m_ member field prefix naming convention.

+++ CorsixTH/Src/run_length_encoder.cpp
@@ -171,7 +171,7 @@ bool IntegerRunLengthEncoder::_areRangesEqual(size_t iObjIdx1, size_t iObjIdx2,
 {
     iObjIdx1 = m_iBufferOffset + iOffset * m_iRecordSize + iObjIdx1 * iObjSize;
     iObjIdx2 = m_iBufferOffset + iOffset * m_iRecordSize + iObjIdx2 * iObjSize;
-    for(int i = 0; i < iObjSize; ++i)
+    for(size_t i = 0; i < iObjSize; ++i)
     {
         if(m_pBuffer[(iObjIdx1 + i) % m_iBufferSize]
         != m_pBuffer[(iObjIdx2 + i) % m_iBufferSize])
